A bunch of refactoring going on here
