+++ resources/readme-images/hac-logo-dark.png
