+++ workshop-1/src/styles/styles.css
@@ -1,0 +1,3 @@
+@import '../styles/App.css';
+@import '../styles/Todo.css';
+@import '../styles/Button.css';
