+++ workshop-2/public/index.html
@@ -1,0 +1,14 @@
+<!doctype html>
+<html lang="en">
+  <head>
+   <meta charset="utf-8">   
+    <link rel="shortcut icon" type="image/x-icon" href="favicon.ico" sizes="32x32"/> 
+    <title>HaC JS Workshop</title>
+  </head>
+  <body>
+    <noscript>
+      You need to enable JavaScript to run this app.
+    </noscript>
+    <div id="root"></div>
+  </body>
+</html>
