+++ workshop-2/src/index.js
@@ -1,0 +1,7 @@
+import React from 'react';
+import ReactDOM from 'react-dom';
+import './styles/styles.css'
+import './styles/index.css';
+import App from './js/App';
+
+ReactDOM.render(<App />, document.getElementById('root'));
