+++ workshop-2/src/styles/App.css
@@ -1,0 +1,55 @@
+.App {
+  text-align: center;
+}
+
+.App_header {
+  padding: 2rem;
+  border-bottom: 1px solid #aaa;
+  margin-bottom: 2rem;
+}
+
+.App_intro {
+  font-size: large;
+}
+
+.App_content {
+  display: flex;
+  justify-content: center;
+  margin: 1rem;
+}
+
+.App_container {
+  max-width: 800px;
+  width: 100%;
+  flex: 0 1 auto;
+  margin-right: 1rem;
+  background-color: #eee;
+  border-radius: 2px;
+  box-shadow: 0 4px 10px rgba(0, 0, 0, 0.2);
+  padding: 1rem;
+}
+
+.App_buttons {
+  display: flex;
+  flex-direction: column;
+  justify-content: left;
+  flex: 0 0 auto;
+}
+
+.App_buttons > * {
+  margin-bottom: 0.5rem;
+}
+
+@media (max-width: 800px) {
+  .App_content {
+    flex-direction: column;
+  }
+
+  .App_buttons {
+    order: -1;
+  }
+
+  .App_container {
+    margin-right: 0;
+  }
+}
