+++ workshop-2/src/styles/index.css
@@ -1,0 +1,7 @@
+body {
+  margin: 0;
+  padding: 0;
+  font-family: Georgia, 'Times New Roman', Times, serif;
+  background: #ddd;
+  color: #333;
+}
