+++ workshop-2/src/styles/styles.css
@@ -1,5 +1,3 @@
 @import '../styles/App.css';
-@import '../styles/Playground.css';
 @import '../styles/Snek.css';
-@import '../styles/Todo.css';
 @import '../styles/Button.css';
