--- workshops/src/styles/Playground.css
@@ -1,9 +1,0 @@
-.Playground {
-  text-align: left;
-}
-
-.Playground-content{
-  padding-left: 0.5em;
-  margin-bottom: 0.75em;
-  margin-top: 0.25em;
-}
